Add the ability to read in metrics_summary.csv from cr outputs, fix bug in experiment dir parser
Now the metrics_summary.csv can be read in (and stored in the mdata object, probably in .uns). I also fixed a recent bug to parse the experiment directory names. I wanted to use regex but unfortunately there's just a few edge cases, specific parsing of the names was easier.