Fixes a bug where wells didn't plot as as dict if it was numbered and not colors
wells can be a dict of well IDs to colors, or anything mappable to colors, but when simply providing floats as values, the function was erroring. intended behavior is to map the floats to the cmap with the default norm, which is now the case. cmap also gets assigned regardless of whether or not color is provided, since we need the cmap to map values to wells if they are not colors.