Update load dataset code for better directory organization
We've recently implemented better organization of our internal directory structures, which is also being impelmented in the way we share data with collaborators and early access customers. In order to make it as easy as possible for collaborators to load data, we've redesigned the directory structure of the internal datasets, adjusted how the 'load' function works, and have implemented a new class SurveyPaths which stores everything together.