Various small enhcancments to io, plotting, and segmentation
In io, we allow the loading in of mtx files now with keyword 'use_mtx', also made slightly more informative error and added 'store_vars' argument to concat_mdatas, which will optionally store the variable names of the concatted adata objects within the mdata (this was useful when we concatted two  datasets each with their own CRISPR guide library; all contained the 'csp' modality but there were unique guides in each dataset that were being dropped without join=outer, and it was useful to know which experiment contained libraries with which guides). In plotting, fix to the set_tick_params function which accounts for every known keyword that might be supplied to set_<x/y>ticklabels. In segmentation, added a (long-needed) undo and redo button; app is now stateful and can undo/redo indefinitely.